add request logic

+++ backend/src/main/java/com/lerhyd/dngame/dao/EntryDao.java
@@ -25,7 +25,7 @@ public interface EntryDao extends JpaRepository<Entry, Long> {
     Entry findById(long id);
 
     @Query("select count(e) from Entry e where e.kira.id =:kiraId and e.pageNumber=:pageNum")
-    int findCntOfEntriesInOnePage(@Param("kiraId") long kiraId, @Param("pageNum") int pageNum);
+    long findCntOfEntriesInOnePage(@Param("kiraId") long kiraId, @Param("pageNum") int pageNum);
 
     boolean existsEntryByVictim_NameAndVictim_SurnameAndVictim_PatronymicAndVictim_Sex(String name,
                                                                                        String surname,
